Added Enemy Collision Behaviours, updated Level Manager and Enemy scripts.
Created a Enemy Collision Behaviour that uses events to determine what happens when something collides with the enemy and calls the TakeDamage function.

Updated the Level Manager and Enemy scripts to accommodate for that.

Fixed a few problems with the ObjectPooler class which now uses gameObject Instance ID's instead of types for searching objects.

Updated Bullet Behaviour script to accommodate for the Enemy Behaivour script.